initial version for optional hidapi usage
